Safari-609.1.11.1/LayoutTests/http/tests/security/contentSecurityPolicy/1.1/plugintypes-invalid-expected.txt

CONSOLE MESSAGE: 'plugin-types' Content Security Policy directive is empty; all plugins will be blocked.

CONSOLE MESSAGE: Refused to load data:application/x-webkit-test-netscape,logifloaded because its MIME type does not appear in the plugin-types directive of the Content Security Policy.
CONSOLE MESSAGE: 'plugin-types' Content Security Policy directive is empty; all plugins will be blocked.

CONSOLE MESSAGE: Refused to load data:application/x-webkit-test-netscape,logifloaded because its MIME type does not appear in the plugin-types directive of the Content Security Policy.
CONSOLE MESSAGE: Invalid plugin type in 'plugin-types' Content Security Policy directive: ''none''.

CONSOLE MESSAGE: Refused to load data:application/x-webkit-test-netscape,logifloaded because its MIME type does not appear in the plugin-types directive of the Content Security Policy.
CONSOLE MESSAGE: Invalid plugin type in 'plugin-types' Content Security Policy directive: 'text'.

CONSOLE MESSAGE: Refused to load data:application/x-webkit-test-netscape,logifloaded because its MIME type does not appear in the plugin-types directive of the Content Security Policy.
CONSOLE MESSAGE: Invalid plugin type in 'plugin-types' Content Security Policy directive: 'text/'.

CONSOLE MESSAGE: Refused to load data:application/x-webkit-test-netscape,logifloaded because its MIME type does not appear in the plugin-types directive of the Content Security Policy.
CONSOLE MESSAGE: Invalid plugin type in 'plugin-types' Content Security Policy directive: '/text'.

CONSOLE MESSAGE: Refused to load data:application/x-webkit-test-netscape,logifloaded because its MIME type does not appear in the plugin-types directive of the Content Security Policy.
CONSOLE MESSAGE: Invalid plugin type in 'plugin-types' Content Security Policy directive: 'text//plain'.

CONSOLE MESSAGE: Refused to load data:application/x-webkit-test-netscape,logifloaded because its MIME type does not appear in the plugin-types directive of the Content Security Policy.
CONSOLE MESSAGE: Invalid plugin type in 'plugin-types' Content Security Policy directive: 'text/plainapplication/nospace'.

CONSOLE MESSAGE: Refused to load data:application/x-webkit-test-netscape,logifloaded because its MIME type does not appear in the plugin-types directive of the Content Security Policy.
CONSOLE MESSAGE: Invalid plugin type in 'plugin-types' Content Security Policy directive: 'text'.

This tests our handling of invalid `plugin-types` CSP directives. Consider this test passing if each of the following frames contains either "PASS" or no text at all.



--------
Frame: '<!--frame1-->'
--------


--------
Frame: '<!--frame2-->'
--------


--------
Frame: '<!--frame3-->'
--------


--------
Frame: '<!--frame4-->'
--------


--------
Frame: '<!--frame5-->'
--------


--------
Frame: '<!--frame6-->'
--------


--------
Frame: '<!--frame7-->'
--------


--------
Frame: '<!--frame8-->'
--------


--------
Frame: '<!--frame9-->'
--------
PASS.
